Update eslint config and separate types
