+++ packages/be-ts-db-lib/package.json
@@ -54,7 +54,6 @@
     "mongodb": "^6.8.0"
   },
   "devDependencies": {
-    "@aneuhold/eslint-config": "^1.0.89",
     "@aneuhold/main-scripts": "^2.1.0",
     "@types/node": "^22.15.3",
     "@types/node-fetch": "^2.6.12",
