+++ packages/be-ts-lib/package.json
@@ -4,7 +4,6 @@
   "license": "MIT",
   "version": "2.0.64",
   "description": "A backend TypeScript library used for common functionality in personal backend projects.",
-  "packageManager": "yarn@4.6.0",
   "type": "module",
   "scripts": {
     "build": "rimraf lib && tsc --project tsconfig.build.json && tsx ./scripts/copyTsFiles.ts",
@@ -50,7 +49,6 @@
     "octokit": "^3.1.2"
   },
   "devDependencies": {
-    "@aneuhold/eslint-config": "^1.0.89",
     "@aneuhold/main-scripts": "^2.0.13",
     "@types/node": "^22.15.3",
     "@types/node-fetch": "^2.6.12",
