+++ packages/core-ts-db-lib/package.json
@@ -4,7 +4,6 @@
   "license": "MIT",
   "version": "2.0.66",
   "description": "A core database library used for personal projects",
-  "packageManager": "yarn@4.6.0",
   "type": "module",
   "scripts": {
     "build": "rimraf lib && tsc --project tsconfig.build.json && tsx ./scripts/copyTsFiles.ts",
@@ -47,7 +46,6 @@
     "bson": "^6.2.0"
   },
   "devDependencies": {
-    "@aneuhold/eslint-config": "^1.0.89",
     "@aneuhold/main-scripts": "^2.0.13",
     "@types/node": "^22.15.3",
     "eslint": "^9.26.0",
