+++ packages/core-ts-lib/src/types/JsonWithVersionProperty.ts
@@ -1,0 +1,3 @@
+export interface JsonWithVersionProperty {
+  version: string;
+}
