+++ packages/core-ts-lib/src/types/PackageJsonInfo.ts
@@ -1,0 +1,6 @@
+import { PackageJson } from './PackageJson.js';
+
+export interface PackageJsonInfo {
+  packageJsonContents: PackageJson;
+  packageJsonPath: string;
+}
