+++ packages/core-ts-lib/src/types/PackageJsonMap.ts
@@ -1,0 +1,5 @@
+import { PackageJsonInfo } from './PackageJsonInfo.js';
+
+export interface PackageJsonMap {
+  [packageName: string]: PackageJsonInfo;
+}
